# book-network/src/main/java/com/crlsistemas/book/feedback/FeedbackResponse.java
package com.crlsistemas.book.feedback;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {
    private Double note;
    private String comment;
    private boolean ownFeedback;
}